refactor(optimizer): move files
